chore: move the mocks folder
